requirements.txt
wget
nemo_toolkit[asr]>=2.dev
git+https://github.com/m-bain/whisperX.git@78dcfaab51005aa703ee21375f81ed31bc248560
#git+https://github.com/adefossez/demucs.git

dora-search>=0.1.12
diffq>=0.2.1
einops
flake8
hydra-colorlog>=1.1
hydra-core>=1.1
julius>=0.2.3
lameenc>=1.2
museval
mypy
openunmix
pyyaml
submitit
torch>=1.8.1
#torchaudio>=0.8,<2.2
torchaudio>=2.2
tqdm
treetable
soundfile>=0.10.3
ipython

git+https://github.com/oliverguhr/deepmultilingualpunctuation.git
git+https://github.com/MahmoudAshraf97/ctc-forced-aligner.git

